api-server: do not reserve the service address
Since 3.2, a service address is reserved in the network which is the
second address of the pool, typically the address ending by `2`. This
breaks the compatibility with previous Contrail versions since a user
is no longer able to use this address.

To preserve the backward compatibitiliy, no service address is
reserved.

Note the `dns_server_address` parameter is used by the Subnet class in
src/config/api-server/vnc_addr_mgmt.py:224 to reserve a service
address.

Change-Id: Ief38a3f14815a1aa7f591e5e986ffd93b04fbb42